+++ generators/app/templates/demo/index.js
@@ -2,9 +2,9 @@
  * <%= ComponentName %> Component Demo for uxcore
  * @author <%= props.authorName %>
  *
- * Copyright 2014-2015, Uxcore Team, Alinw.
+ * Copyright 2015-2016, Uxcore Team, Alinw.
  * All rights reserved.
  */
 
  var Demo = require('./<%= ComponentName %>Demo');
- React.render(<Demo/>, document.getElementById('UXCoreDemo'));
+ ReactDOM.render(<Demo/>, document.getElementById('UXCoreDemo'));